+++ src/field/focused_laser.rs
@@ -271,6 +271,7 @@ impl Field for FocusedLaser {
                 u_prime: u + (n as f64) * kappa - k,
                 pol,
                 a_eff: a,
+                chi: a * (u * kappa),
                 absorption: (n as f64) * kappa[0],
             };
             Some(event)
@@ -295,6 +296,7 @@ impl Field for FocusedLaser {
                 u_p: q_p,
                 frac: 1.0 / rate_increase,
                 a_eff: a,
+                chi: a * (ell * kappa),
                 absorption: (n as f64) * kappa[0],
             };
             (prob, pol_new, Some(event))