+++ src/main.rs
@@ -22,9 +22,9 @@ use hdf5_writer;
 #[cfg(feature = "hdf5-output")]
 unzip_n::unzip_n!(pub 7);
 #[cfg(feature = "hdf5-output")]
-unzip_n::unzip_n!(pub 8);
-#[cfg(feature = "hdf5-output")]
 unzip_n::unzip_n!(pub 9);
+#[cfg(feature = "hdf5-output")]
+unzip_n::unzip_n!(pub 10);
 
 mod constants;
 mod field;
@@ -131,6 +131,7 @@ fn collide<F: Field, R: Rng>(field: &F, incident: Particle, rng: &mut R, current
                         *current_id = *current_id + 1;
                         let photon = Particle::create(Species::Photon, r)
                             .with_payload(event.a_eff)
+                            .with_parent_chi(event.chi)
                             .with_weight(pt.weight())
                             .with_id(id)
                             .with_parent_id(pt.id())
@@ -175,6 +176,7 @@ fn collide<F: Field, R: Rng>(field: &F, incident: Particle, rng: &mut R, current
                             .with_weight(event.frac * pt.weight())
                             .with_id(id)
                             .with_payload(event.a_eff)
+                            .with_parent_chi(event.chi)
                             .with_parent_id(pt.id())
                             .update_absorbed_energy(0.5 * event.absorption)
                             .with_normalized_momentum(event.u_e);
@@ -182,6 +184,7 @@ fn collide<F: Field, R: Rng>(field: &F, incident: Particle, rng: &mut R, current
                             .with_weight(event.frac * pt.weight())
                             .with_id(id + 1)
                             .with_payload(event.a_eff)
+                            .with_parent_chi(event.chi)
                             .with_parent_id(pt.id())
                             .update_absorbed_energy(0.5 * event.absorption)
                             .with_normalized_momentum(event.u_p);
@@ -1084,7 +1087,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                 // Write particle data
                 let fs = file.new_group("final-state")?;
 
-                let (x, p, pol, w, a, n, id, pid) = photons
+                let (x, p, pol, w, a, chi, n, id, pid) = photons
                     .iter()
                     .map(|pt| (
                         pt.position().convert(&units.length),
@@ -1092,6 +1095,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                         pt.polarization(),
                         pt.weight(),
                         pt.payload(),
+                        pt.parent_chi(),
                         pt.interaction_count(),
                         pt.id(),
                         pt.parent_id()
@@ -1110,6 +1114,10 @@ fn main() -> Result<(), Box<dyn Error>> {
                         .with_desc("normalized amplitude (RMS under LMA) at point of emission")?
                         .with_alias("xi")?
                         .write(&a[..])?
+                    .new_dataset("parent_chi")?
+                        .with_unit("1")?
+                        .with_desc("quantum parameter (RMS under LMA) of parent at point of emission")?
+                        .write(&chi[..])?
                     .new_dataset("n_pos")?
                         .with_unit("1")?
                         .with_desc("total probability of pair creation for the photon")?
@@ -1177,7 +1185,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                         .with_desc("four-momentum of the electron")?
                         .write(&p[..])?;
 
-                let (x, x0, p, w, n, abs, id, pid, a) = positrons
+                let (x, x0, p, w, n, abs, id, pid, a, chi) = positrons
                     .iter()
                     .map(|pt| (
                         pt.position().convert(&units.length),
@@ -1188,7 +1196,8 @@ fn main() -> Result<(), Box<dyn Error>> {
                         pt.absorbed_energy().convert(&units.energy),
                         pt.id(),
                         pt.parent_id(),
-                        pt.payload()
+                        pt.payload(),
+                        pt.parent_chi(),
                     ))
                     .unzip_n_vec();
 
@@ -1204,6 +1213,10 @@ fn main() -> Result<(), Box<dyn Error>> {
                         .with_desc("normalized amplitude (RMS under LMA) at point of creation")?
                         .with_alias("xi")?
                         .write(&a[..])?
+                    .new_dataset("parent_chi")?
+                        .with_unit("1")?
+                        .with_desc("quantum parameter (RMS under LMA) of parent at point of creation")?
+                        .write(&chi[..])?
                     .new_dataset("n_gamma")?
                         .with_unit("1")?
                         .with_desc("total number of photons emitted by the positron")?
@@ -1234,7 +1247,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                 if keep_decayed_photons {
                     let is = file.new_group("intermediate-state")?;
 
-                    let (x, p, pol, w, a, n, id, pid) = decayed_photons
+                    let (x, p, pol, w, a, chi, n, id, pid) = decayed_photons
                         .iter()
                         .map(|pt| (
                             pt.position().convert(&units.length),
@@ -1243,6 +1256,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                             pt.polarization(),
                             pt.weight(),
                             pt.payload(),
+                            pt.parent_chi(),
                             pt.interaction_count(),
                             pt.id(),
                             pt.parent_id()
@@ -1261,6 +1275,10 @@ fn main() -> Result<(), Box<dyn Error>> {
                             .with_desc("normalized amplitude (RMS under LMA) at point of emission")?
                             .with_alias("xi")?
                             .write(&a[..])?
+                        .new_dataset("parent_chi")?
+                            .with_unit("1")?
+                            .with_desc("quantum parameter (RMS under LMA) of parent at point of emission")?
+                            .write(&chi[..])?
                         .new_dataset("n_pos")?
                             .with_unit("1")?
                             .with_desc("total probability of pair creation for the photon")?